recognize timezone 'CST' correctly in China
On some platforms, for example, Linux will use CST (China Standard Time)
as time string suffix of GMT+8, hence the external log tables in may be
parsed incorrectly for users in China. Issue #3574.

I tracked that change in Postgres to this commit firstly:

   commit e8dac6f1e09e5bc7604caa6f111d6c80a8b98295
   Author: Tom Lane <tgl@sss.pgh.pa.us>
   Date:   Wed Dec 24 16:35:23 2014 -0500

       Add CST (China Standard Time) to our lists of timezone abbreviations.

       For some reason this seems to have been missed when the lists in
       src/timezone/tznames/ were first constructed.  We can't put it in Default
       because of the conflict with US CST, but we should certainly list it among
       the alternative entries in Asia.txt.  (I checked for other oversights, but
       all the other abbreviations that are in current use according to the IANA
       files seem to be accounted for.)  Noted while responding to bug #12326.

Then add a standard timezone 'Asia' in OVERRIDE mode, and expose it in default list.
With it, we may "set timezone_abbreviations = 'Asia'" to recognize CST naturally,
and also correctly.